+++ Source/EbsdLib/LaueOps/CubicOps.cpp
@@ -58,9 +58,9 @@ namespace CubicHigh
 {
 
 static const std::array<size_t, 3> OdfNumBins = {18, 18, 18}; // Represents a 5Deg bin
-static const std::array<double, 3> OdfDimInitValue = {std::pow((0.75f * (EbsdLib::Constants::k_PiOver4 - sinf(EbsdLib::Constants::k_PiOver4))), (1.0f / 3.0)),
-                                                      std::pow((0.75f * (EbsdLib::Constants::k_PiOver4 - sinf(EbsdLib::Constants::k_PiOver4))), (1.0f / 3.0)),
-                                                      std::pow((0.75f * (EbsdLib::Constants::k_PiOver4 - sinf(EbsdLib::Constants::k_PiOver4))), (1.0f / 3.0))};
+static const std::array<double, 3> OdfDimInitValue = {std::pow((0.75 * (EbsdLib::Constants::k_PiOver4 - std::sin(EbsdLib::Constants::k_PiOver4))), (1.0 / 3.0)),
+                                                      std::pow((0.75 * (EbsdLib::Constants::k_PiOver4 - std::sin(EbsdLib::Constants::k_PiOver4))), (1.0 / 3.0)),
+                                                      std::pow((0.75 * (EbsdLib::Constants::k_PiOver4 - std::sin(EbsdLib::Constants::k_PiOver4))), (1.0 / 3.0))};
 
 static const std::array<double, 3> OdfDimStepValue = {OdfDimInitValue[0] / static_cast<double>(OdfNumBins[0] / 2), OdfDimInitValue[1] / static_cast<double>(OdfNumBins[1] / 2),
                                                       OdfDimInitValue[2] / static_cast<double>(OdfNumBins[2] / 2)};
@@ -1721,7 +1721,7 @@ std::vector<EbsdLib::UInt8ArrayType::Pointer> CubicOps::generatePoleFigure(PoleF
   if(config.labels.size() > 1) { label1 = config.labels.at(1); }
   if(config.labels.size() > 2) { label2 = config.labels.at(2); }
 
-  int numOrientations = config.eulers->getNumberOfTuples();
+  size_t numOrientations = config.eulers->getNumberOfTuples();
 
   // Create an Array to hold the XYZ Coordinates which are the coords on the sphere.
   // this is size for CUBIC ONLY, <001> Family
@@ -2364,7 +2364,7 @@ EbsdLib::UInt8ArrayType::Pointer CubicOps::generateMisorientationTriangleLegend(
     refQuat.w() = 1;
     //have rodrigues vector, need quat
     double tanAng = sqrt(d1[i] * d1[i] + d2[i] * d2[i] + d3[i] * d3[i]);
-    double cosAng = cosf(atanf(tanAng));
+    double cosAng = std::cos(atanf(tanAng));
     quat.x() = d1[i] * cosAng * tanAng;
     quat.y() = d2[i] * cosAng * tanAng;
     quat.z(), () = d3[i] * cosAng * tanAng;
